+++ packages/riverpod/lib/src/providers/stream_provider.dart
@@ -188,17 +188,6 @@ class $StreamProviderElement<StateT>
     }
     super.onError(value, seamless: seamless);
   }
-
-  @override
-  bool updateShouldNotify(
-    AsyncValue<StateT> previous,
-    AsyncValue<StateT> next,
-  ) {
-    return FutureModifierElement.handleUpdateShouldNotify(
-      previous,
-      next,
-    );
-  }
 }
 
 /// The [Family] of a [StreamProvider]